Add capabilities to read exr images
This should offer better support than what CImg can do natively.

Especially, an extension of file path can be added, with an additional /
after the pather followed by Exr.Layer.Name.exrlayer or
Exr.channel.name.exrchannel to load a specific layer or channel as an
image.

This is enabled only when libstevi is linked with OpenEXR, which is done
automatically when cmake find OpenEXR.