Ensure libstevi is built as a shared library
This is to avoid conflicts if other targets trying to build without open
MP try to use libstevi in release mode. If the lib is build as a static
library, then the target needs openMP libraries without openMP, which
complicate the configuration a lot.

It is also cleaner for other things.