Order aware ransac sampling strategy
For a lot of problems in computer vision, it can be guessed with
reasonable efficency, that some measures are much more likely to be
outliers. This commit add a Ransac Strategy which samples the measures
in order, assuming the first measures are more likely to be inliers (the
end user is responsible to order the measures accordingly when using the
strategy).